修改

+++ WriteDocument.aspx.cs
@@ -1442,17 +1442,24 @@ namespace WebApplication1
                             }
                             else
                             {
-                                Lbl_Eorr.Visible = true;
-                                using (SqlConnection cn = new SqlConnection(tmpdbhelper.DB_CnStr))
+                                string LvlLast = ((TextBox)GridView2.Rows[GridView2.Rows.Count - 1].FindControl("Txt_Lvl")).Text.Trim();
+                                string EIDLast = ((TextBox)GridView2.Rows[GridView2.Rows.Count - 1].FindControl("Txt_EID")).Text.Trim();
+                                if (EIDLast != "" && LvlLast != "")
                                 {
-                                    cn.Open();
-                                    SqlCommand cmddelete = new SqlCommand("Delete From Detail Where SID=@SID");
-                                    cmddelete.Connection = cn;
-                                    cmddelete.Parameters.AddWithValue("@SID",Lbl_SID.Text);
-                                    cmddelete.ExecuteNonQuery();
-                                    cn.Close();
-                                }
+                                    Lbl_Eorr.Visible = true;
+                                    using (SqlConnection cn = new SqlConnection(tmpdbhelper.DB_CnStr))
+                                    {
+                                        cn.Open();
+                                        SqlCommand cmddelete = new SqlCommand("Delete From Detail Where SID=@SID");
+                                        cmddelete.Connection = cn;
+                                        cmddelete.Parameters.AddWithValue("@SID", Lbl_SID.Text);
+                                        cmddelete.ExecuteNonQuery();
+                                        cn.Close();
+                                    }
                                     return;
+                                }
+
+
                             }
                         }
